+++ R/globals.R
@@ -1,5 +1,5 @@
 ## globals for ivd function
-utils::globalVariables(c('N', 'K', 'Kr', 'X', 'groupid', 'Z', 'S', 'Sr', 'zeta', 'X_scale', 'Z_scale', 'J', 'P', 'sigma_rand', 'L', 'z', 'ss'))
+utils::globalVariables(c('N', 'K', 'Kr', 'X', 'groupid', 'Z', 'S', 'Sr', 'zeta', 'X_scale', 'Z_scale', 'J', 'P', 'sigma_rand', 'L', 'z', 'ss', 'student', 'nu_raw'))
 ## globabls for plot.ivd function
 utils::globalVariables(c('pip', 'id', 'label', 'WAIC', 'zscore', 'geom_text_repel'))
 ## globals for plot.pip_sensitivity